repositories/AttackVLA/SpatialVLA/robogcg_spatialvla/requirements.txt
absl-py==2.1.0
accelerate==1.2.1
annotated-types==0.7.0
array-record==0.5.1
astunparse==1.6.3
attrs==24.2.0
bddl==3.5.0
cachetools==5.5.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
dm-tree==0.1.8
docker-pycreds==0.4.0
draccus==0.8.0
easydict==1.13
einops==0.8.0
etils==1.11.0
evdev==1.7.1
exceptiongroup==1.2.2
fastjsonschema==2.21.1
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gitdb==4.0.11
gitpython==3.1.43
glfw==2.8.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.68.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.12.1
huggingface-hub==0.26.5
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib-resources==6.4.5
iniconfig==2.0.0
joblib==1.4.2
json-numpy==2.1.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-core==5.7.2
jupytext==1.16.4
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.43.0
markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mergedeep==1.3.4
mink==0.0.5
ml-dtypes==0.2.0
mujoco==3.2.6
mypy-extensions==1.0.0
nbformat==5.10.4
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
openexr==3.3.2
opt-einsum==3.4.0
packaging==24.2
peft==0.11.1
platformdirs==4.3.6
pluggy==1.5.0
promise==2.3
protobuf==3.20.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.3
pydantic-core==2.27.1
pygments==2.18.0
pynput==1.7.7
pyopengl==3.1.7
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-xlib==0.33
pyyaml-include==1.4.1
qpsolvers==4.4.0
quadprog==0.1.13
referencing==0.35.1
regex==2024.11.6
requests-oauthlib==2.0.0
rich==13.9.4
robosuite==1.4.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.1.99
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
timm==0.9.10
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.1
trimesh==4.5.3
triton==2.2.0
typeguard==2.13.3
typing-extensions==4.12.2
typing-inspect==0.9.0
wandb==0.19.1
werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0